requirements.txt
Django==3.2.6
flake8==3.9.2
flake8-builtins==1.5.3
flake8-django==1.1.2
flake8-import-order==0.18.1
flake8-print==4.0.0
djangorestframework==3.12.4
dj-rest-auth==2.1.11
django-allauth==0.44.0
djangorestframework-simplejwt==4.7.2
